flow-api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
supabase==2.7.0
slowapi==0.1.9
python-multipart==0.0.12
prometheus-fastapi-instrumentator==7.0.0
pydantic-settings==2.4.0
pytest==8.3.0
pytest-asyncio==0.23.0
google-generativeai>=0.8.0
openai==1.45.0
python-dotenv==1.0.0
Pillow==10.4.0
requests==2.32.0
PyJWT==2.8.0
